Added 'pinned' tasks for quick access
Cron now clears on non-recurring tasks

Deleting orphaned tasks now excludes those referred to in routine steps

Fix for cleanup popup on RoutineStep completion

Fixes for RoutineStep shifting

Started processing for splitting NetDuration into separate thread for multiprocessing

+++ src/main/java/com/trajan/negentropy/server/backend/util/OrphanTaskCleaner.java
@@ -1,6 +1,8 @@
 package com.trajan.negentropy.server.backend.util;
 
+import com.trajan.negentropy.model.entity.routine.QRoutineStepEntity;
 import com.trajan.negentropy.server.backend.EntityQueryService;
+import com.trajan.negentropy.server.backend.repository.RoutineStepRepository;
 import com.trajan.negentropy.server.backend.repository.TaskRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,14 +15,19 @@ import org.springframework.transaction.annotation.Transactional;
 public class OrphanTaskCleaner {
     @Autowired private EntityQueryService entityQueryService;
     @Autowired private TaskRepository taskRepository;
+    @Autowired private RoutineStepRepository stepRepository;
 
     public void deleteAllOrphanedTasks() {
         entityQueryService.findOrphanedTasks().forEach(task -> {
-            log.debug("Deleting orphaned task: " + task);
-            try {
-                taskRepository.delete(task);
-            } catch (Throwable t) {
-                log.error("Error deleting orphaned task: " + task, t);
+            if (!stepRepository.exists(QRoutineStepEntity.routineStepEntity.task.eq(task))) {
+                log.debug("Deleting orphaned task <" + task + ">");
+                try {
+                    taskRepository.delete(task);
+                } catch (Throwable t) {
+                    log.error("Error deleting orphaned task: " + task, t);
+                }
+            } else {
+                log.debug("Skipping orphaned task <" + task.name() + "> because it is referred to by routine steps");
             }
         });
     }
